Move projectSelectionList platform entities to the list package
